% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ALL_HALOCARBON_CONCENTRATIONS}
\alias{ALL_HALOCARBON_CONCENTRATIONS}
\title{Get all of the possible Hector halocarbon concentrations}
\usage{
ALL_HALOCARBON_CONCENTRATIONS()
}
\value{
Character vector of variable names.
}
\description{
This function returns a vector of the possible hector halocarbon concentrations
that can be accessed with \code{\link{fetchvars}}.
}
\examples{
\dontrun{
ini <- system.file(package = "hector", "input/hector_ssp245.ini")
hc <- newcore(ini)
run(hc)
out <- fetchvars(core = hc, dates = 1900:2100, vars = ALL_HALOCARBON_CONCENTRATIONS())
print(out)
}
}
\seealso{
Other outputs: 
\code{\link{ALL_HALOCARBON_RF}()},
\code{\link{ALL_VARS}()}
}
\concept{outputs}
