% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ALL_HALOCARBON_EMISSIONS}
\alias{ALL_HALOCARBON_EMISSIONS}
\title{Get all of the possible Hector halocarbon emissions}
\usage{
ALL_HALOCARBON_EMISSIONS()
}
\value{
Character vector of variable names.
}
\description{
This function returns a vector of the possible hector halocarbon emissions
that can be accessed with \code{\link{fetchvars}} or set with \code{\link{setvar}}.
}
\examples{
\dontrun{
ini <- system.file(package = "hector", "input/hector_ssp245.ini")
hc <- newcore(ini)
run(hc)
out <- fetchvars(core = hc, dates = 1900:2100, vars = ALL_HALOCARBON_EMISSIONS())
print(out)
}
}
\concept{haloemiss}
