% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ALL_HALOCARBON_RF}
\alias{ALL_HALOCARBON_RF}
\title{Get all of the possible Hector halocarbon radiative forcings}
\usage{
ALL_HALOCARBON_RF()
}
\value{
Character vector of variable names.
}
\description{
This function returns a vector of the possible hector halocarbon forcings
that can be accessed with \code{\link{fetchvars}}.
}
\examples{
\dontrun{
ini <- system.file(package = "hector", "input/hector_ssp245.ini")
hc <- newcore(ini)
run(hc)
out <- fetchvars(core = hc, dates = 1900:2100, vars = ALL_HALOCARBON_RF())
print(out)
}
}
\seealso{
Other outputs: 
\code{\link{ALL_HALOCARBON_CONCENTRATIONS}()},
\code{\link{ALL_VARS}()}
}
\concept{outputs}
