% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ALL_VARS}
\alias{ALL_VARS}
\title{Get all of the possible Hector output variables}
\usage{
ALL_VARS()
}
\value{
Character vector of variable names.
}
\description{
This function returns all of the possible hector output variables
that can be accessed with \code{\link{fetchvars}}.
}
\examples{
\dontrun{
ini <- system.file(package = "hector", "input/hector_ssp245.ini")
hc <- newcore(ini)
run(hc)
out <- fetchvars(core = hc, dates = 1900:2100, vars = ALL_VARS())
print(out)
}
}
\seealso{
Other outputs: 
\code{\link{ALL_HALOCARBON_CONCENTRATIONS}()},
\code{\link{ALL_HALOCARBON_RF}()}
}
\concept{outputs}
