% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_create_biome}
\alias{c_create_biome}
\title{Create a biome}
\usage{
c_create_biome(core, biome)
}
\arguments{
\item{core}{Handle to the Hector instance that is to be run.}

\item{biome}{(character) Name of new biome}
}
\description{
Create a biome
}
