% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_delete_biome}
\alias{c_delete_biome}
\title{Delete a biome}
\usage{
c_delete_biome(core, biome)
}
\arguments{
\item{core}{Handle to the Hector instance that is to be run.}

\item{biome}{(character) Name of biome to delete}
}
\description{
Delete a biome
}
