% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{create_biome}
\alias{create_biome}
\title{Create new biome}
\usage{
create_biome(core, biome, veg_c0, detritus_c0, soil_c0, npp_flux0,
  warmingfactor = 1, beta = 0.36)
}
\arguments{
\item{core}{Hector core}

\item{biome}{Name of new biome}

\item{veg_c0}{Initial vegetation C pool}

\item{detritus_c0}{Initial detritus C pool}

\item{soil_c0}{Initial soil C pool}

\item{npp_flux0}{Initial net primary productivity}

\item{warmingfactor}{Temperature multiplier (default =
`1.0`)}

\item{beta}{CO2 fertilization effect (default = `0.36`)}
}
\value{
Hector core, invisibly
}
\description{
Create new biome
}
\author{
Alexey Shiklomanov
}
