% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{cum_vars_input}
\alias{cum_vars_input}
\title{Cumulate exogenous Hector input variables}
\usage{
cum_vars_input(core, dates, lib_funcs)
}
\arguments{
\item{core}{Hector core object}

\item{dates}{Vector of dates to fetch}
}
\value{
rslt_tot Dataframe of variables
}
\description{
This function cumulates all available exogenous input variables for the active
Hector core, such as emissions and pre-industrial species concentrations.
}
\seealso{
Other fetchvars_all helper function: \code{\link{cum_vars_output}},
  \code{\link{cum_vars_params}}
}
\concept{fetchvars_all helper function}
