% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{cum_vars_output}
\alias{cum_vars_output}
\title{Cumulate Hector output variables}
\usage{
cum_vars_output(core, dates, lib_funcs)
}
\arguments{
\item{core}{Hector core object}

\item{dates}{Vector of dates}
}
\value{
rslt_tot Dataframe containing the variables
}
\description{
This function cumulates all available output variables from an active Hector
core (e.g.,  Concentrations, Forcings, etc.)
}
\seealso{
Other fetchvars_all helper function: \code{\link{cum_vars_input}},
  \code{\link{cum_vars_params}}
}
\concept{fetchvars_all helper function}
