% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{cum_vars_params}
\alias{cum_vars_params}
\title{Cumulate Hector parameters}
\usage{
cum_vars_params(core)
}
\arguments{
\item{core}{Hector object}
}
\value{
rslt_tot Dataframe containing the parameters/variables
}
\description{
This function cumulates all available Hector parameters, such as BETA, ESC,
and Volcanic Scale
}
\seealso{
Other fetchvars_all helper function: \code{\link{cum_vars_input}},
  \code{\link{cum_vars_output}}
}
\concept{fetchvars_all helper function}
