% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delete_biome_impl}
\alias{delete_biome_impl}
\title{Delete a biome}
\usage{
delete_biome_impl(core, biome)
}
\arguments{
\item{core}{Handle to the Hector instance that is to be run.}

\item{biome}{(character) Name of biome to delete}
}
\description{
Delete a biome
}
