% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{DELTA_CO2}
\alias{DELTA_CO2}
\alias{RF_N2O}
\alias{DELTA_N2O}
\alias{DELTA_CF4}
\alias{DELTA_C2F6}
\alias{DELTA_HFC23}
\alias{DELTA_HFC32}
\alias{DELTA_HFC4310}
\alias{DELTA_HFC125}
\alias{DELTA_HFC134A}
\alias{DELTA_HFC143A}
\alias{DELTA_HFC227EA}
\alias{DELTA_HFC245FA}
\alias{DELTA_SF6}
\alias{DELTA_CFC11}
\alias{DELTA_CFC12}
\alias{DELTA_CFC113}
\alias{DELTA_CFC114}
\alias{DELTA_CFC115}
\alias{DELTA_CCL4}
\alias{DELTA_CH3CCL3}
\alias{DELTA_HCFC22}
\alias{DELTA_HCFC141B}
\alias{DELTA_HCFC142B}
\alias{DELTA_HALON1211}
\alias{DELTA_HALON1301}
\alias{DELTA_HALON2402}
\alias{DELTA_CH3CL}
\alias{DELTA_CH3BR}
\alias{DELTA_CH4}
\alias{delta}
\title{Identifiers for Hector forcing component parameters}
\usage{
DELTA_CO2()

RF_N2O()

DELTA_N2O()

DELTA_CF4()

DELTA_C2F6()

DELTA_HFC23()

DELTA_HFC32()

DELTA_HFC4310()

DELTA_HFC125()

DELTA_HFC134A()

DELTA_HFC143A()

DELTA_HFC227EA()

DELTA_HFC245FA()

DELTA_SF6()

DELTA_CFC11()

DELTA_CFC12()

DELTA_CFC113()

DELTA_CFC114()

DELTA_CFC115()

DELTA_CCL4()

DELTA_CH3CCL3()

DELTA_HCFC22()

DELTA_HCFC141B()

DELTA_HCFC142B()

DELTA_HALON1211()

DELTA_HALON1301()

DELTA_HALON2402()

DELTA_CH3CL()

DELTA_CH3BR()

DELTA_CH4()
}
\description{
These identifiers specify the tropospheric adjustments for the
stratospheric-temperature adjusted radiative forcings.
These values must be a number between -1 and 1, and
can be read and/or set by hectors forcing and halocarbon components.
}
\section{Functions}{
\itemize{
\item \code{DELTA_CO2()}: the foring tropospheric adjustments for N2O

\item \code{RF_N2O()}: the foring tropospheric adjustments for N2O

\item \code{DELTA_N2O()}: Radiative forcing due to N2O

\item \code{DELTA_CF4()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_C2F6()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC23()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC32()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing '
@export

\item \code{DELTA_HFC4310()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC125()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC134A()}: the tropospheric adjustments used to convert from tratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC143A()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC227EA()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HFC245FA()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_SF6()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CFC11()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CFC12()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CFC113()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CFC114()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CFC115()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CCL4()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CH3CCL3()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HCFC22()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HCFC141B()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HCFC142B()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HALON1211()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HALON1301()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_HALON2402()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CH3CL()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CH3BR()}: the tropospheric adjustments used to convert from stratospheric-temperature adjusted radiative forcing to effective radiative forcing

\item \code{DELTA_CH4()}: Radiative forcing tropospheric adjustment for CH4

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{forcings} forcing
values provided from the hector forcing component.

Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
