% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{fetchvars_all}
\alias{fetchvars_all}
\title{Fetch all available variables from a running Hector core}
\usage{
fetchvars_all(core, dates = NULL, scenario = NULL)
}
\arguments{
\item{core}{Hector core object}

\item{dates}{Vector of dates, optional. Defines the time span to retrieve timeseries
variables from. If not provided, the Hector core object's start and end years will be used.}

\item{scenario}{Optional scenario name. If not specified, the name of the Hector
core object will be used.}
}
\value{
Dataframe containing all Hector variables
}
\description{
This function is similar to fetchvars, except it fetches all available
Hector variables, including exogenous inputs (e.g., emissions & pre-industrial
concentrations), parameters (e.g., alpha, beta, ECS), and output
variables (e.g., concentrations & forcings). Variables with and without
timeseries capabilities are supported.
}
\seealso{
Other main user interface functions: \code{\link{fetchvars}},
  \code{\link{newcore}}, \code{\link{reset}},
  \code{\link{run}}, \code{\link{setvar}},
  \code{\link{shutdown}}
}
\concept{main user interface functions}
