% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{RF_TOTAL}
\alias{RF_TOTAL}
\alias{RF_ALBEDO}
\alias{RF_CO2}
\alias{RF_H2O_STRAT}
\alias{RF_O3_TROP}
\alias{RF_BC}
\alias{RF_OC}
\alias{RF_NH3}
\alias{RF_SO2}
\alias{RF_ACI}
\alias{RF_VOL}
\alias{RF_MISC}
\alias{Q_CO2}
\alias{RF_CH4}
\alias{forcings}
\title{Identifiers for capabilities in the Hector forcing component}
\usage{
RF_TOTAL()

RF_ALBEDO()

RF_CO2()

RF_H2O_STRAT()

RF_O3_TROP()

RF_BC()

RF_OC()

RF_NH3()

RF_SO2()

RF_ACI()

RF_VOL()

RF_MISC()

Q_CO2()

RF_CH4()
}
\description{
These identifiers specify forcing values that can be provided by hector via
the forcing component. All of the values corresponding to these identifiers
are read-only (\emph{i.e.}, they can only appear in \code{\link{GETDATA}}
messages.)
}
\section{Functions}{
\itemize{
\item \code{RF_TOTAL()}: Total radiative forcing

\item \code{RF_ALBEDO()}: Radiative forcing due to albedo

\item \code{RF_CO2()}: Radiative forcing due to CO2

\item \code{RF_H2O_STRAT()}: Radiative forcing due to stratospheric water vapor

\item \code{RF_O3_TROP()}: Radiative forcing due to tropospheric ozone

\item \code{RF_BC()}: Radiative forcing due to black carbon aerosol-radiation interactions

\item \code{RF_OC()}: Radiative forcing due to organic carbon aerosol-radiation interactions

\item \code{RF_NH3()}: Radiative forcing due to NH3 aerosol-radiation interactions

\item \code{RF_SO2()}: Radiative forcing due to SO2 aerosol-radiation interactions

\item \code{RF_ACI()}: aerosol-cloud interactions, includes contributions from SO2, BC and OC

\item \code{RF_VOL()}: Radiative forcing due to volcanic activity

\item \code{RF_MISC()}: Radiative miscellaneous forcing read in from ini.

\item \code{Q_CO2()}: Radiative forcing efficiency for doubling of CO2

\item \code{RF_CH4()}: Radiative forcing due to methane

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{rho} for
parameters related to radiative forcing efficiency.

Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
