% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\docType{data}
\name{fxntable}
\alias{fxntable}
\title{A data.frame containing information R hector function name and hector string names.}
\format{
A data.frame of 2 columns and 239 rows.
\describe{
 \item{fxn}{String of the R hector function name}
 \item{string}{String of variable name}
}
}
\usage{
fxntable
}
\description{
A data.frame containing information R hector function name and hector string names.
}
\keyword{datasets}
