% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{get_biome_inits}
\alias{get_biome_inits}
\title{Retrieve the initial conditions and parameters for a given biome}
\usage{
get_biome_inits(core, biome)
}
\arguments{
\item{core}{Hector core}

\item{biome}{Name of biome for which to retrieve data}
}
\value{
Named numeric vector of biome initial conditions and parameters
}
\description{
Internal helper function for biome functions.
}
\author{
Alexey Shiklomanov
}
