% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_biome_list}
\alias{get_biome_list}
\title{Retrieve the current list of biomes for a Hector instance}
\usage{
get_biome_list(core)
}
\arguments{
\item{core}{Handle to the Hector instance from which to retrieve
the biome list.}
}
\description{
Retrieve the current list of biomes for a Hector instance
}
