% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getdate}
\alias{getdate}
\title{Get the current date for a Hector instance}
\usage{
getdate(core)
}
\arguments{
\item{core}{Handle to a Hector instance}
}
\value{
The current date in the Hector instance
}
\description{
The "current date" is the last year that the Hector instance has completed.
}
