% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{getfxn}
\alias{getfxn}
\title{Get the hector function associated with a specific string}
\usage{
getfxn(str)
}
\arguments{
\item{strstr}{String name to find the Hector function for}
}
\value{
Hector function name
}
\description{
This function provides the Hector function that returns a the Hector string
}
\examples{
getfxn("beta")
getfxn("q10_rh")
}
\seealso{
getunits
}
