% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getunits}
\alias{getunits}
\title{Get the hector unit strings for one or more variables}
\usage{
getunits(vars)
}
\arguments{
\item{vars}{Capability strings for variables to get units for.}
}
\value{
Character vector of unit strings. Variables not found will have \code{NA} entries.
}
\description{
This function provides units only for variables that can be set in the model
using \code{\link{setvar}}.  Variables read from the model with
\code{\link{fetchvars}} will have a units column in the data frame they are
returned in.
}
