% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{CF4_CONSTRAIN}
\alias{CF4_CONSTRAIN}
\alias{C2F6_CONSTRAIN}
\alias{HFC23_CONSTRAIN}
\alias{HFC32_CONSTRAIN}
\alias{HFC4310_CONSTRAIN}
\alias{HFC125_CONSTRAIN}
\alias{HFC134A_CONSTRAIN}
\alias{HFC143A_CONSTRAIN}
\alias{HFC227EA_CONSTRAIN}
\alias{HFC245FA_CONSTRAIN}
\alias{SF6_CONSTRAIN}
\alias{CFC11_CONSTRAIN}
\alias{CFC12_CONSTRAIN}
\alias{CFC113_CONSTRAIN}
\alias{CFC114_CONSTRAIN}
\alias{CFC115_CONSTRAIN}
\alias{CCL4_CONSTRAIN}
\alias{CH3CCL3_CONSTRAIN}
\alias{HCFC22_CONSTRAIN}
\alias{HCFC141B_CONSTRAIN}
\alias{HCFC142B_CONSTRAIN}
\alias{HALON1211_CONSTRAIN}
\alias{HALON1301_CONSTRAIN}
\alias{HALON2402_CONSTRAIN}
\alias{CH3CL_CONSTRAIN}
\alias{CH3BR_CONSTRAIN}
\alias{haloconstrain}
\title{Identifiers for halocarbon concentration constraints}
\usage{
CF4_CONSTRAIN()

C2F6_CONSTRAIN()

HFC23_CONSTRAIN()

HFC32_CONSTRAIN()

HFC4310_CONSTRAIN()

HFC125_CONSTRAIN()

HFC134A_CONSTRAIN()

HFC143A_CONSTRAIN()

HFC227EA_CONSTRAIN()

HFC245FA_CONSTRAIN()

SF6_CONSTRAIN()

CFC11_CONSTRAIN()

CFC12_CONSTRAIN()

CFC113_CONSTRAIN()

CFC114_CONSTRAIN()

CFC115_CONSTRAIN()

CCL4_CONSTRAIN()

CH3CCL3_CONSTRAIN()

HCFC22_CONSTRAIN()

HCFC141B_CONSTRAIN()

HCFC142B_CONSTRAIN()

HALON1211_CONSTRAIN()

HALON1301_CONSTRAIN()

HALON2402_CONSTRAIN()

CH3CL_CONSTRAIN()

CH3BR_CONSTRAIN()
}
\description{
These identifiers correspond to concentration constraints for halocarbons. In
all cases, the expected input units are volumetric parts per trillion
(\code{"ppvt"}).
}
\section{Functions}{
\itemize{
\item \code{CF4_CONSTRAIN()}: Concentration constraint for CF4

\item \code{C2F6_CONSTRAIN()}: Concentration constraint for C2F6

\item \code{HFC23_CONSTRAIN()}: Concentration constraint for HFC-23

\item \code{HFC32_CONSTRAIN()}: Concentration constraint for HFC-32

\item \code{HFC4310_CONSTRAIN()}: Concentration constraint for HFC-4310

\item \code{HFC125_CONSTRAIN()}: Concentration constraint for HFC-125

\item \code{HFC134A_CONSTRAIN()}: Concentration constraint for HFC-134a

\item \code{HFC143A_CONSTRAIN()}: Concentration constraint for HFC-143a

\item \code{HFC227EA_CONSTRAIN()}: Concentration constraint for HFC-227ea

\item \code{HFC245FA_CONSTRAIN()}: Concentration constraint for HFC-254fa

\item \code{SF6_CONSTRAIN()}: Concentration constraint for sulfur hexafluoride

\item \code{CFC11_CONSTRAIN()}: Concentration constraint for CFC-11

\item \code{CFC12_CONSTRAIN()}: Concentration constraint for CFC-12

\item \code{CFC113_CONSTRAIN()}: Concentration constraint for CFC-113

\item \code{CFC114_CONSTRAIN()}: Concentration constraint for CFC-114

\item \code{CFC115_CONSTRAIN()}: Concentration constraint for CFC-115

\item \code{CCL4_CONSTRAIN()}: Concentration constraint for carbon tetrachloride

\item \code{CH3CCL3_CONSTRAIN()}: Concentration constraint for trichloroethane

\item \code{HCFC22_CONSTRAIN()}: Concentration constraint for HCFC-22

\item \code{HCFC141B_CONSTRAIN()}: Concentration constraint for HCFC-141b

\item \code{HCFC142B_CONSTRAIN()}: Concentration constraint for HCFC-142b

\item \code{HALON1211_CONSTRAIN()}: Concentration constraint for halon-1211

\item \code{HALON1301_CONSTRAIN()}: Concentration constraint for halon-1301

\item \code{HALON2402_CONSTRAIN()}: Concentration constraint for halon-2402

\item \code{CH3CL_CONSTRAIN()}: Concentration constraint for chloromethane

\item \code{CH3BR_CONSTRAIN()}: Concentration constraint for bromomethane

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers:
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
