% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{EMISSIONS_CF4}
\alias{EMISSIONS_CF4}
\alias{EMISSIONS_C2F6}
\alias{EMISSIONS_HFC23}
\alias{EMISSIONS_HFC32}
\alias{EMISSIONS_HFC4310}
\alias{EMISSIONS_HFC125}
\alias{EMISSIONS_HFC134A}
\alias{EMISSIONS_HFC143A}
\alias{EMISSIONS_HFC227EA}
\alias{EMISSIONS_HFC245FA}
\alias{EMISSIONS_SF6}
\alias{EMISSIONS_CFC11}
\alias{EMISSIONS_CFC12}
\alias{EMISSIONS_CFC113}
\alias{EMISSIONS_CFC114}
\alias{EMISSIONS_CFC115}
\alias{EMISSIONS_CCL4}
\alias{EMISSIONS_CH3CCL3}
\alias{EMISSIONS_HCFC22}
\alias{EMISSIONS_HCFC141B}
\alias{EMISSIONS_HCFC142B}
\alias{EMISSIONS_HALON1211}
\alias{EMISSIONS_HALON1301}
\alias{EMISSIONS_HALON2402}
\alias{EMISSIONS_CH3CL}
\alias{EMISSIONS_CH3BR}
\alias{haloemiss}
\title{Identifiers for halocarbon emissions}
\usage{
EMISSIONS_CF4()

EMISSIONS_C2F6()

EMISSIONS_HFC23()

EMISSIONS_HFC32()

EMISSIONS_HFC4310()

EMISSIONS_HFC125()

EMISSIONS_HFC134A()

EMISSIONS_HFC143A()

EMISSIONS_HFC227EA()

EMISSIONS_HFC245FA()

EMISSIONS_SF6()

EMISSIONS_CFC11()

EMISSIONS_CFC12()

EMISSIONS_CFC113()

EMISSIONS_CFC114()

EMISSIONS_CFC115()

EMISSIONS_CCL4()

EMISSIONS_CH3CCL3()

EMISSIONS_HCFC22()

EMISSIONS_HCFC141B()

EMISSIONS_HCFC142B()

EMISSIONS_HALON1211()

EMISSIONS_HALON1301()

EMISSIONS_HALON2402()

EMISSIONS_CH3CL()

EMISSIONS_CH3BR()
}
\description{
These identifiers correspond to emissions values for halocarbons. They are
(for now) write only, meaning they can only appear in \code{\link{SETDATA}}
messages. In all cases, the expected input units are gigagrams
(\code{"Gg"}).
}
\section{Functions}{
\itemize{
\item \code{EMISSIONS_CF4()}: Emissions for CF4

\item \code{EMISSIONS_C2F6()}: Emissions for C2F6

\item \code{EMISSIONS_HFC23()}: Emissions for HFC-23

\item \code{EMISSIONS_HFC32()}: Emissions for HFC-32

\item \code{EMISSIONS_HFC4310()}: Emissions for HFC-4310

\item \code{EMISSIONS_HFC125()}: Emissions for HFC-125

\item \code{EMISSIONS_HFC134A()}: Emissions for HFC-134a

\item \code{EMISSIONS_HFC143A()}: Emissions for HFC-143a

\item \code{EMISSIONS_HFC227EA()}: Emissions for HFC-227ea

\item \code{EMISSIONS_HFC245FA()}: Emissions for HFC-254fa

\item \code{EMISSIONS_SF6()}: Emissions for sulfur hexafluoride

\item \code{EMISSIONS_CFC11()}: Emissions for CFC-11

\item \code{EMISSIONS_CFC12()}: Emissions for CFC-12

\item \code{EMISSIONS_CFC113()}: Emissions for CFC-113

\item \code{EMISSIONS_CFC114()}: Emissions for CFC-114

\item \code{EMISSIONS_CFC115()}: Emissions for CFC-115

\item \code{EMISSIONS_CCL4()}: Emissions for carbon tetrachloride

\item \code{EMISSIONS_CH3CCL3()}: Emissions for trichloroethane

\item \code{EMISSIONS_HCFC22()}: Emissions for HCFC-22

\item \code{EMISSIONS_HCFC141B()}: Emissions for HcFC-141b

\item \code{EMISSIONS_HCFC142B()}: Emissions for HCFC-142b

\item \code{EMISSIONS_HALON1211()}: Emissions for halon-1211

\item \code{EMISSIONS_HALON1301()}: Emissions for halon-1301

\item \code{EMISSIONS_HALON2402()}: Emissions for halon-2402

\item \code{EMISSIONS_CH3CL()}: Emissions for chloromethane

\item \code{EMISSIONS_CH3BR()}: Emissions for bromomethane

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
