% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{RF_CF4}
\alias{RF_CF4}
\alias{RF_C2F6}
\alias{RF_HFC23}
\alias{RF_HFC32}
\alias{RF_HFC4310}
\alias{RF_HFC125}
\alias{RF_HFC134A}
\alias{RF_HFC143A}
\alias{RF_HFC227EA}
\alias{RF_HFC245FA}
\alias{RF_SF6}
\alias{RF_CFC11}
\alias{RF_CFC12}
\alias{RF_CFC113}
\alias{RF_CFC114}
\alias{RF_CFC115}
\alias{RF_CCL4}
\alias{RF_CH3CCL3}
\alias{RF_HCFC22}
\alias{RF_HCFC141B}
\alias{RF_HCFC142B}
\alias{RF_HALON1211}
\alias{RF_HALON1301}
\alias{RF_HALON2402}
\alias{RF_CH3CL}
\alias{RF_CH3BR}
\alias{haloforcings}
\title{Identifiers for halocarbon forcings}
\usage{
RF_CF4()

RF_C2F6()

RF_HFC23()

RF_HFC32()

RF_HFC4310()

RF_HFC125()

RF_HFC134A()

RF_HFC143A()

RF_HFC227EA()

RF_HFC245FA()

RF_SF6()

RF_CFC11()

RF_CFC12()

RF_CFC113()

RF_CFC114()

RF_CFC115()

RF_CCL4()

RF_CH3CCL3()

RF_HCFC22()

RF_HCFC141B()

RF_HCFC142B()

RF_HALON1211()

RF_HALON1301()

RF_HALON2402()

RF_CH3CL()

RF_CH3BR()
}
\description{
These identifiers specify forcing values that can be provided by hector via
one of the myriad halocarbon components. All of the values corresponding to
these identifiers are read-only (\emph{i.e.}, they can only appear in
\code{\link{GETDATA}} messages.)  The forcings returned are the
\emph{relative} forcings, with the base year (typically 1750) values
subtracted off.
}
\section{Functions}{
\itemize{
\item \code{RF_CF4()}: Radiative forcing due to CF4

\item \code{RF_C2F6()}: Radiative forcing due to C2F6

\item \code{RF_HFC23()}: Radiative forcing due to HFC-23

\item \code{RF_HFC32()}: Radiative forcing due to HFC-32

\item \code{RF_HFC4310()}: Radiative forcing due to HFC-4310

\item \code{RF_HFC125()}: Radiative forcing due to HFC-125

\item \code{RF_HFC134A()}: Radiative forcing due to HFC-134a

\item \code{RF_HFC143A()}: Radiative forcing due to HFC-143a

\item \code{RF_HFC227EA()}: Radiative forcing due to HFC227ea

\item \code{RF_HFC245FA()}: Radiative forcing due to HFC-245fa

\item \code{RF_SF6()}: Radiative forcing due to sulfur hexafluoride

\item \code{RF_CFC11()}: Radiative forcing due to CFC-11

\item \code{RF_CFC12()}: Radiative forcing due to CFC-12

\item \code{RF_CFC113()}: Radiative forcing due to CFC-113

\item \code{RF_CFC114()}: Radiative forcing due to CFC-114

\item \code{RF_CFC115()}: Radiative forcing due to CFC-115

\item \code{RF_CCL4()}: Radiative forcing due to carbon tetrachloride

\item \code{RF_CH3CCL3()}: Radiative forcing due to trichloroethane

\item \code{RF_HCFC22()}: Radiative forcing due to HCFC-22

\item \code{RF_HCFC141B()}: Radiative forcing due to HCFC-141b

\item \code{RF_HCFC142B()}: Radiative forcing due to HCCFC-142b

\item \code{RF_HALON1211()}: Radiative forcing due to halon-1211

\item \code{RF_HALON1301()}: Radiative forcing due to halon-1301

\item \code{RF_HALON2402()}: Radiative forcing due to halon-2402

\item \code{RF_CH3CL()}: Radiative forcing due to chloromethane

\item \code{RF_CH3BR()}: Radiative forcing due to bromomethane

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{forcings} for forcings from other sources.

Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
