% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R, R/hector.R
\name{getdate}
\alias{getdate}
\alias{hectorutil}
\alias{isactive}
\alias{startdate}
\alias{enddate}
\alias{getname}
\title{\strong{getdate}: Get the current date for a Hector instance}
\usage{
getdate(core)

isactive(core)

startdate(core)

enddate(core)

getname(core)
}
\arguments{
\item{core}{Hector instance to operate on}
}
\description{
\strong{getdate}: Get the current date for a Hector instance

Utility functions for Hector instances

\strong{isactive}: Indicate whether a Hector instance is active

\strong{startdate}: Report the start date for a Hector instance

\strong{enddate}: Report the end date for a Hector instance

\strong{getname}: Report the name for a Hector instance
}
