% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{keyvalue}
\alias{keyvalue}
\title{Convert a vector of strings of the form \code{"name=value"} to a named R
list}
\usage{
keyvalue(x)
}
\arguments{
\item{x}{Character vector of the form \code{"name=value"}}
}
\value{
Named list
}
\description{
Convert a vector of strings of the form \code{"name=value"} to a named R
list
}
\author{
Alexey Shiklomanov
}
