% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{merge_biomes}
\alias{merge_biomes}
\title{Merge multiple biomes into a single biome}
\usage{
merge_biomes(core, old_biomes, new_biome, ...)
}
\arguments{
\item{core}{Hector core}

\item{old_biomes}{Names of biomes to be merged}

\item{new_biome}{Name of new merged biome}

\item{...}{Additional parameters for the new biome, as set by
[create_biome()].}
}
\description{
Merge multiple biomes into a single biome
}
\author{
Alexey Shiklomanov
}
