% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{methods}
\alias{methods}
\alias{format.hcore}
\alias{print.hcore}
\title{Hector instance object methods}
\usage{
\method{format}{hcore}(x, ...)

\method{print}{hcore}(x, ...)
}
\arguments{
\item{x}{Hector core object}

\item{...}{Additional arguments}
}
\description{
Hector instance object methods
}
\section{Methods (by generic)}{
\itemize{
\item \code{format}: Format method

\item \code{print}: Print method
}}

