% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{msgtype}
\alias{msgtype}
\alias{GETDATA}
\alias{SETDATA}
\alias{BIOME_SPLIT_CHAR}
\title{Message type identifiers}
\usage{
GETDATA()

SETDATA()

BIOME_SPLIT_CHAR()
}
\description{
These identifiers indicate the type of a message so that the core can route
it to the correct method in the recipient.
}
\section{Functions}{
\itemize{
\item \code{GETDATA}: Message type for retrieving data from a component

\item \code{SETDATA}: Message type for setting data in a component

\item \code{BIOME_SPLIT_CHAR}: Character used to separate biome from variable name
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\keyword{internal}
