% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{newcore}
\alias{newcore}
\title{Create and initialize a new hector instance}
\usage{
newcore(
  inifile,
  loglevel = 0,
  suppresslogging = TRUE,
  name = "Unnamed Hector core"
)
}
\arguments{
\item{inifile}{(String) name of the hector input file.}

\item{loglevel}{(int) minimum message level to output in logs (see \code{\link{loglevels}}).}

\item{suppresslogging}{(bool) If true, suppress all logging (loglevel is ignored in this case).}

\item{name}{(string) An optional name to identify the core.}
}
\value{
handle for the Hector instance.
}
\description{
The object returned is a handle to the newly created instance.  It will be required as an
argument for all functions that operate on the instance.  Creating multiple instances
simultaneously is supported.
}
\seealso{
Other main user interface functions:
\code{\link[=fetchvars]{fetchvars()}},
\code{\link[=get_tracking_data]{get_tracking_data()}},
\code{\link[=reset]{reset()}},
\code{\link[=run]{run()}},
\code{\link[=setvar]{setvar()}},
\code{\link[=shutdown]{shutdown()}}
}
\concept{main user interface functions}
