% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{normalize_paths}
\alias{normalize_paths}
\title{Apply Hector internal logic to convert relative paths to absolute paths}
\usage{
normalize_paths(value_list, base_dir = NULL)
}
\arguments{
\item{value_list}{Named list of values from a particular section}

\item{base_dir}{Base directory to use if files are not found}
}
\value{
\code{value_list}, with all path variables replaced with
normalized, absolute paths.
}
\description{
Apply Hector internal logic to convert relative paths to absolute paths
}
