% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{TRACKING_DATE}
\alias{TRACKING_DATE}
\alias{PREINDUSTRIAL_CO2}
\alias{BETA}
\alias{Q10_RH}
\alias{WARMINGFACTOR}
\alias{F_NPPV}
\alias{F_NPPD}
\alias{F_LITTERD}
\alias{ECS}
\alias{AERO_SCALE}
\alias{VOLCANIC_SCALE}
\alias{LO_WARMING_RATIO}
\alias{DIFFUSIVITY}
\alias{parameters}
\title{Identifiers for model parameters}
\usage{
TRACKING_DATE()

PREINDUSTRIAL_CO2()

BETA(biome = "")

Q10_RH(biome = "")

WARMINGFACTOR(biome = "")

F_NPPV(biome = "")

F_NPPD(biome = "")

F_LITTERD(biome = "")

ECS()

AERO_SCALE()

VOLCANIC_SCALE()

LO_WARMING_RATIO()

DIFFUSIVITY()
}
\arguments{
\item{biome}{Biome for which to retrieve parameter. If missing or `""`, default to `"global"`.}
}
\description{
These identifiers correspond to settable parameters that change the model
behavior and are subject to uncertainty. All of these can be set using the
\code{\link{SETDATA}} message type. Changing any of these parameters will
typically invalidate the hector core's internal state; therefore, after
setting one or more of these values you should call \code{\link{reset}} before
attempting to run the model again. This will rerun the spinup and produce a
new internally consistent state. Attempting to run the model without resetting
first will usually produce an error (often with a message about failing to
conserve mass).
}
\section{Functions}{
\itemize{
\item \code{TRACKING_DATE()}: Start of carbon tracking (Year)

\item \code{PREINDUSTRIAL_CO2()}: Preindustrial CO2 concentration (\code{"ppmv CO2"})

\item \code{BETA()}: CO2 fertilization factor (\code{"(unitless)"})

\item \code{Q10_RH()}: Heterotrophic respiration temperature sensitivity factor (\code{"(unitless)"})

\item \code{WARMINGFACTOR()}: Biome-specific warming factor (`(unitless)`)

\item \code{F_NPPV()}: NPP fraction to vegetation (\code{"(unitless)"})

\item \code{F_NPPD()}: NPP fraction to detritus (\code{"(unitless)"})

\item \code{F_LITTERD()}: Litter fraction to detritus (\code{"(unitless)"})

\item \code{ECS()}: Equilibrium Climate Sensitivity (\code{"degC"})

\item \code{AERO_SCALE()}: Aerosol forcing scaling factor (\code{"(unitless)"})

\item \code{VOLCANIC_SCALE()}: Volcanic forcing scaling factor (\code{"(unitless)"})

\item \code{LO_WARMING_RATIO()}: Land-Ocean Warming Ratio (\code{"(unitless)"}), by default set to 0 meaning that the land ocean warming ratio is an emergent property of Hector's temperature component otherwise the user defined land ocean warming ratio will be used.

\item \code{DIFFUSIVITY()}: Ocean heat diffusivity (\code{"cm2/s"})

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}}
}
\concept{capability identifiers}
