% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{parse_timeseries}
\alias{parse_timeseries}
\title{From a list of values, convert the ones that are time series
(e.g.\code{"Ftalbedo[1980]"}) to a \code{data.frame}. Leave scalar and
character values (including \emph{paths} to time series data) as is.}
\usage{
parse_timeseries(value_list)
}
\arguments{
\item{value_list}{Named list of values from a particular section}
}
\value{
\code{value_list}, but with all time series combined into
\code{data.frame}s and moved to the end of the list.
}
\description{
From a list of values, convert the ones that are time series
(e.g.\code{"Ftalbedo[1980]"}) to a \code{data.frame}. Leave scalar and
character values (including \emph{paths} to time series data) as is.
}
\author{
Alexey Shiklomanov
}
