% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{read_ini}
\alias{read_ini}
\title{Read Hector INI file into list}
\usage{
read_ini(file)
}
\arguments{
\item{file}{Path to INI file}
}
\description{
Read Hector INI file into list
}
\examples{
rcp45_ini <- read_ini(
  system.file("input", "hector_rcp45.ini", package = "hector")
)
rcp45_ini$core$run_name
rcp45_ini$onelineocean$enabled
}
\author{
Alexey Shiklomanov
}
