% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rename_biome}
\alias{rename_biome}
\title{Rename an existing biome}
\usage{
rename_biome(core, oldname, newname)
}
\arguments{
\item{core}{Handle to the Hector instance that is to be run.}

\item{oldname}{(character) Name of existing biome to be replaced}

\item{newname}{(character) Name of new biome}
}
\description{
This will create a new biome called `newname`, assign it all of
the C stocks and parameter values from biome `oldname`, and delete
biome `oldname`.
}
