% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Run the Hector climate model}
\usage{
run(core, runtodate = -1)
}
\arguments{
\item{core}{Handle to the Hector instance that is to be run.}

\item{runtodate}{Date to run to.  The default is to run to the end date
configured in the input file used to initialize the core.}
}
\value{
The Hector instance handle
}
\description{
Run Hector up through the specified time.  This function does not return the
results of the run.  To get results, run \code{fetch}.
}
\seealso{
Other main user interface functions:
\code{\link[=fetchvars]{fetchvars()}},
\code{\link[=get_tracking_data]{get_tracking_data()}},
\code{\link[=newcore]{newcore()}},
\code{\link[=reset]{reset()}},
\code{\link[=setvar]{setvar()}},
\code{\link[=shutdown]{shutdown()}}
}
\concept{main user interface functions}
