% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{runscenario}
\alias{runscenario}
\title{Run a single scenario}
\usage{
runscenario(infile)
}
\arguments{
\item{infile}{INI-format file containing the scenario definition}
}
\value{
Data frame containing Hector output for default variables
}
\description{
Run the scenario defined by the input file and return a data frame containing
results for the default variable list.  The default variable list can be
changed by setting the \code{hector.default.fetchvars} option, as described
in \code{\link{fetchvars}}.
}
