% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param.R
\name{set_param}
\alias{set_param}
\title{Set parameter inside Hector core}
\usage{
set_param(core, parameter, value, unit, reset_date = 0)
}
\arguments{
\item{core}{Hector core object}

\item{parameter}{Name of parameter to set. Must be length 1. See \link{parameters}.}

\item{value}{(numeric) New parameter value. Must be length 1 (scalar).}

\item{unit}{(character) Parameter unit. See \link{parameters}.}

\item{reset_date}{(numeric) Year to reset to. Default is 0.0. See
\code{date} argument of \code{\link[=reset]{reset()}}.}
}
\value{
(invisibly) Hector core with parameter changed
}
\description{
Set a parameter inside a Hector core to a particular value, and
then \code{\link[=reset]{reset()}} the core.
}
\author{
Alexey Shiklomanov
}
