% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shutdown}
\alias{shutdown}
\title{Shut down a hector instance}
\usage{
shutdown(core)
}
\arguments{
\item{core}{Handle to a Hector instance}
}
\value{
The Hector instance handle
}
\description{
Shutting down an instance will free the instance itself and all of the
objects it created. Any attempted operation on the instance after that will
raise an error.
}
\section{Caution}{

This function should be called as \code{mycore <- shutdown(mycore)} so that
the change from active to inactive will be recorded in the caller.
}

\seealso{
Other main user interface functions:
\code{\link[=fetchvars]{fetchvars()}},
\code{\link[=get_tracking_data]{get_tracking_data()}},
\code{\link[=newcore]{newcore()}},
\code{\link[=reset]{reset()}},
\code{\link[=run]{run()}},
\code{\link[=setvar]{setvar()}}
}
\concept{main user interface functions}
