% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{split_biome}
\alias{split_biome}
\title{Create a new biome by splitting an existing one}
\usage{
split_biome(core, biome, from_biome = get_biome_list(core)[[1]],
  fveg_c = 0.5, fdetritus_c = fveg_c, fsoil_c = fveg_c,
  fnpp_flux0 = f_vegc, beta = NULL, warmingfactor = NULL)
}
\description{
Create a new biome by splitting an existing one
}
