% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{GLOBAL_TAS}
\alias{GLOBAL_TAS}
\alias{GMST}
\alias{SST}
\alias{OCEAN_TAS}
\alias{LAND_TAS}
\alias{FLUX_MIXED}
\alias{FLUX_INTERIOR}
\alias{HEAT_FLUX}
\alias{temperature}
\title{Identifiers for variables associated with the temperature component}
\usage{
GLOBAL_TAS()

GMST()

SST()

OCEAN_TAS()

LAND_TAS()

FLUX_MIXED()

FLUX_INTERIOR()

HEAT_FLUX()
}
\description{
All of these variables can be read using the \code{\link{GETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{GLOBAL_TAS()}: Global mean air temperature anomaly

\item \code{GMST()}: Global mean surface temperature anomaly

\item \code{SST()}: Average sea surface temperature anomaly

\item \code{OCEAN_TAS()}: Average air temperature anomaly over the ocean

\item \code{LAND_TAS()}: Average air temperature anomaly over land, land surface temperature and air temperature over land are assumed to be equivalent.

\item \code{FLUX_MIXED()}: Heat flux into the mixed layer of the ocean

\item \code{FLUX_INTERIOR()}: Heat flux into the interior layer of the ocean

\item \code{HEAT_FLUX()}: Total heat flux into the ocean

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_BC]{EMISSIONS_BC()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
