% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{try_numeric}
\alias{try_numeric}
\title{Try to convert a value to numeric, and if it throws a warning,
return the original value.}
\usage{
try_numeric(x)
}
\arguments{
\item{x}{Value to be converted to numeric}
}
\value{
\code{as.numeric(x)} if successful, or \code{x} if not
}
\description{
Try to convert a value to numeric, and if it throws a warning,
return the original value.
}
\author{
Alexey Shiklomanov
}
