% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\docType{data}
\name{unitstable}
\alias{unitstable}
\title{A data.frame containing information on all of the Hector variable created by units-data.R
used within the getunits function.}
\format{
A data.frame of 2 columns and 114 rows.
\describe{
 \item{variable}{String of the hector variable}
 \item{units}{String of the unit value}
}
}
\usage{
unitstable
}
\description{
A data.frame containing information on all of the Hector variable created by units-data.R
used within the getunits function.
}
\keyword{datasets}
