% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{use_biomes}
\alias{use_biomes}
\title{Enable biomes}
\usage{
use_biomes(core, first_biome = "biome1")
}
\arguments{
\item{first_biome}{(Character) Name of first biome. Default is `"biome1"`.}
}
\value{
Hector core, invisibly
}
\description{
Enable biomes
}
\author{
Alexey Shiklomanov
}
