% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ini.R
\name{write_ini}
\alias{write_ini}
\title{Write R list to Hector INI file}
\usage{
write_ini(ini_list, file)
}
\arguments{
\item{ini_list}{Hector INI named, nested list, such as that
returned by \code{\link[=read_ini]{read_ini()}}.}

\item{file}{Output file, as character string or connection. See
\code{\link[base:writeLines]{base::writeLines()}}.}
}
\value{
\code{ini_list}, invisibly
}
\description{
Write R list to Hector INI file
}
\examples{
l <- list(a = list(a1 = 5, a2 = 3.5),
          b = list(some_file = "hello", `b3[50]` = 5))
tmp <- tempfile()
write_ini(l, tmp)
# Write to `stdout()`
write_ini(l, "")
}
\author{
Alexey Shiklomanov
}
