% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{ALL_VARS}
\alias{ALL_VARS}
\title{Get all of the possible Hector output variables}
\usage{
ALL_VARS()
}
\value{
Character vector of variable names.
}
\description{

}
\examples{
\dontrun{
ini <- system.file(package = "hector", "input/hector_ssp245.ini")
hc <- newcore(ini)
run(hc)
out <- fetchvars(core = hc, dates = 1900:2100, vars = ALL_VARS())
print(out)
}
}
\concept{outputs}
