% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{NBP}
\alias{NBP}
\alias{CONCENTRATIONS_CO2}
\alias{NPP}
\alias{RH}
\alias{ATMOSPHERIC_CO2}
\alias{FFI_EMISSIONS}
\alias{DACCS_UPTAKE}
\alias{LUC_EMISSIONS}
\alias{LUC_UPTAKE}
\alias{CO2_CONSTRAIN}
\alias{NBP_CONSTRAIN}
\alias{VEG_C}
\alias{DETRITUS_C}
\alias{SOIL_C}
\alias{PERMAFROST_C}
\alias{EARTH_C}
\alias{NPP_FLUX0}
\alias{carboncycle}
\title{Identifiers for variables in the Hector carbon cycle component}
\usage{
NBP()

CONCENTRATIONS_CO2()

NPP()

RH()

ATMOSPHERIC_CO2()

FFI_EMISSIONS()

DACCS_UPTAKE()

LUC_EMISSIONS()

LUC_UPTAKE()

CO2_CONSTRAIN()

NBP_CONSTRAIN()

VEG_C(biome = "")

DETRITUS_C(biome = "")

SOIL_C(biome = "")

PERMAFROST_C(biome = "")

EARTH_C()

NPP_FLUX0(biome = "")
}
\arguments{
\item{biome}{Name of biome (leave empty for global)}
}
\description{
These identifiers correspond to variables that can be read and/or set in the
carbon cycle
}
\section{Functions}{
\itemize{
\item \code{CO2_CONSTRAIN()}: Constrain atmospheric CO2 concentration (\code{"(ppmv CO2)"})

\item \code{NBP_CONSTRAIN()}: Constrain net biome production (land-atmosphere flux) (\code{"(PgC/yr)"})

\item \code{VEG_C()}: Vegetation C pool (`"Pg C"`)

\item \code{DETRITUS_C()}: Vegetation detritus C pool (`"Pg C"`)

\item \code{SOIL_C()}: Soil C pool (`"Pg C"`)

\item \code{PERMAFROST_C()}: Permafrost C pool (`"Pg C"`)

\item \code{NPP_FLUX0()}: Initial net primary productivity (NPP)flux (`"Pg C year^-1"`)

}}
\section{Output variables}{

These variables can be read using the \code{\link{GETDATA}} message type:
\describe{
\item{CONCENTRATIONS_CO2}{Atmospheric CO2 concentration}
\item{ATMOSPHERIC_CO2}{Atmospheric CO2 in units of C}
\item{NBP}{Net biome production. Annual global C flux from atmosphere into
  the land. A positive value means a net flux from atmosphere into land
  (i.e. land is a net carbon sink), while a negative value means a net flux
  from land into the atmosphere (i.e. land is a net carbon source). Note
  that despite the name--which follows disciplinary convention, see Chapin
  et al. 2006--this is currently a globally-averaged variable.}
\item{FFI_EMISSIONS}{Fossil fuel and industrial emissions}
\item{LUC_EMISSIONS}{Land use change emissions}
\item{NPP}{Net primary production}
\item{RH}{Heterotrophic respiration}
\item{EARTH_C}{Earth pool}
}
}

\section{Input variables}{

\describe{
\item{FFI_EMISSIONS}{Fossil fuel and industrial emissions (\code{"Pg C/yr"})}
\item{LUC_EMISSIONS}{Land use change emissions (\code{"Pg C/yr"})}
\item{CO2_CONSTRAIN}{Prescribed atmospheric CO2 concentration (\code{"ppmv CO2"})}
\item{NBP_CONSTRAIN}{Prescribed net biome production (land-atmosphere C flux) (\code{"Pg C/yr"})}
\item{DACCS_UPTAKE}{Direct air carbon capture and storage}
\item{LUC_UPTAKE}{Land use change uptake}
}
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{concentrations} for other gas concentrations and
\link{emissions} for other gas emissions

Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_CH4}()},
\code{\link{CONCENTRATIONS_N2O}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{EMISSIONS_SO2}()},
\code{\link{FTOT_CONSTRAIN}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{OCEAN_UPTAKE}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
