% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{CONCENTRATIONS_N2O}
\alias{CONCENTRATIONS_N2O}
\alias{PREINDUSTRIAL_N2O}
\alias{PREINDUSTRIAL_O3}
\alias{CONCENTRATIONS_O3}
\alias{concentrations}
\title{Identifiers for miscellaneous concentrations not elsewhere described}
\usage{
CONCENTRATIONS_N2O()

PREINDUSTRIAL_N2O()

PREINDUSTRIAL_O3()

CONCENTRATIONS_O3()
}
\description{
All of these variables can be read using the \code{\link{GETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{CONCENTRATIONS_N2O}: Atmospheric N2O concentration

\item \code{PREINDUSTRIAL_N2O}: Preindustrial atmospheric N2O concentration

\item \code{PREINDUSTRIAL_O3}: Preindustrial ozone concentration

\item \code{CONCENTRATIONS_O3}: Atmospheric ozone concentration
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_CH4}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{EMISSIONS_SO2}()},
\code{\link{FTOT_CONSTRAIN}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{NBP}()},
\code{\link{OCEAN_UPTAKE}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
