% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{concentrations}
\alias{concentrations}
\alias{ATMOSPHERIC_N2O}
\alias{PREINDUSTRIAL_N2O}
\alias{PREINDUSTRIAL_O3}
\alias{ATMOSPHERIC_O3}
\title{Identifiers for miscellaneous concentrations not elsewhere described}
\usage{
ATMOSPHERIC_N2O()

PREINDUSTRIAL_N2O()

PREINDUSTRIAL_O3()

ATMOSPHERIC_O3()
}
\description{
All of these variables can be read using the \code{\link{GETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{ATMOSPHERIC_N2O}: Atmospheric N2O concentration

\item \code{PREINDUSTRIAL_N2O}: Preindustrial atmospheric N2O concentration

\item \code{PREINDUSTRIAL_O3}: Preindustrial ozone concentration

\item \code{ATMOSPHERIC_O3}: Atmospheric ozone concentration
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{emissions}}, \code{\link{forcings}},
  \code{\link{haloemiss}}, \code{\link{haloforcings}},
  \code{\link{methane}}, \code{\link{ocean}},
  \code{\link{parameters}}, \code{\link{so2}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
