% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{FTOT_CONSTRAIN}
\alias{FTOT_CONSTRAIN}
\alias{CH4_CONSTRAIN}
\alias{N2O_CONSTRAIN}
\alias{TAS_CONSTRAIN}
\alias{constraints}
\title{Identifiers for constraints}
\usage{
FTOT_CONSTRAIN()

CH4_CONSTRAIN()

N2O_CONSTRAIN()

TAS_CONSTRAIN()
}
\description{
These identifiers correspond to Hector's constraint capabilities
}
\section{Functions}{
\itemize{
\item \code{FTOT_CONSTRAIN}: Constrain total radiative forcing

\item \code{CH4_CONSTRAIN}: CH4 concentration constraint \code{"ppbv CH4"}

\item \code{N2O_CONSTRAIN}: N2O concentration constraint (\code{"pbbv N2O"})

\item \code{TAS_CONSTRAIN}: Constrain global mean temperature  (\code{"(degC)"})
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
haloconstrains

Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_CH4}()},
\code{\link{CONCENTRATIONS_N2O}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{EMISSIONS_SO2}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{NBP}()},
\code{\link{OCEAN_UPTAKE}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
