% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{constraints}
\alias{constraints}
\alias{FTOT_CONSTRAIN}
\alias{CH4_CONSTRAIN}
\alias{N2O_CONSTRAIN}
\alias{TAS_CONSTRAIN}
\title{Identifiers for constraints}
\usage{
FTOT_CONSTRAIN()

CH4_CONSTRAIN()

N2O_CONSTRAIN()

TAS_CONSTRAIN()
}
\description{
These identifiers correspond to Hector's constraint capabilities
}
\section{Functions}{
\itemize{
\item \code{FTOT_CONSTRAIN}: Constrain total radiative forcing

\item \code{CH4_CONSTRAIN}: CH4 concentration constraint \code{"ppbv CH4"}

\item \code{N2O_CONSTRAIN}: N2O concentration constraint (\code{"pbbv N2O"})

\item \code{TAS_CONSTRAIN}: Constrain global mean temperature  (\code{"(degC)"})
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
haloconstrains

Other capability identifiers: 
\code{\link{carboncycle}},
\code{\link{concentrations}},
\code{\link{delta}},
\code{\link{emissions}},
\code{\link{forcings}},
\code{\link{haloconstrain}},
\code{\link{haloemiss}},
\code{\link{haloforcings}},
\code{\link{methane}},
\code{\link{ocean}},
\code{\link{parameters}},
\code{\link{rho}},
\code{\link{so2}},
\code{\link{temperature}}
}
\concept{capability identifiers}
