% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{delta}
\alias{delta}
\alias{DELTA_CO2}
\alias{RF_N2O}
\alias{DELTA_N2O}
\alias{DELTA_CH4}
\title{Identifiers for Hector forcing component parameters}
\usage{
DELTA_CO2()

RF_N2O()

DELTA_N2O()

DELTA_CH4()
}
\description{
These identifiers specify the tropospheric adjustments for the
stratospheric-temperature adjusted radiative forcings.
These values must be a number between -1 and 1, and
can be read and/or set by hectors forcing and halocarbon components.
}
\section{Functions}{
\itemize{
\item \code{DELTA_CO2()}: the foring tropospheric adjustments for N2O

\item \code{RF_N2O()}: the foring tropospheric adjustments for N2O

\item \code{DELTA_N2O()}: Radiative forcing due to N2O

\item \code{DELTA_CH4()}: Radiative forcing tropospheric adjustment for CH4

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{forcings}
forcing values provided from the hector forcing component.

Other capability identifiers: 
\code{\link{carboncycle}},
\code{\link{concentrations}},
\code{\link{constraints}},
\code{\link{emissions}},
\code{\link{forcings}},
\code{\link{haloconstrain}},
\code{\link{haloemiss}},
\code{\link{haloforcings}},
\code{\link{methane}},
\code{\link{ocean}},
\code{\link{parameters}},
\code{\link{rho}},
\code{\link{so2}},
\code{\link{temperature}}
}
\concept{capability identifiers}
