% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{EMISSIONS_BC}
\alias{EMISSIONS_BC}
\alias{EMISSIONS_N2O}
\alias{NAT_EMISSIONS_N2O}
\alias{EMISSIONS_NOX}
\alias{EMISSIONS_CO}
\alias{EMISSIONS_NMVOC}
\alias{EMISSIONS_OC}
\alias{EMISSIONS_NH3}
\alias{emissions}
\title{Identifiers for miscellaneous emissions not elsewhere described}
\usage{
EMISSIONS_BC()

EMISSIONS_N2O()

NAT_EMISSIONS_N2O()

EMISSIONS_NOX()

EMISSIONS_CO()

EMISSIONS_NMVOC()

EMISSIONS_OC()

EMISSIONS_NH3()
}
\description{
All of these variables can be set using the \code{\link{SETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{EMISSIONS_BC()}: Black carbon emissions

\item \code{EMISSIONS_N2O()}: N2O emissions (\code{"Tg N"})

\item \code{NAT_EMISSIONS_N2O()}: Natural N2O emissions (\code{"Tg N"})

\item \code{EMISSIONS_NOX()}: Emissions for NOx compounds (\code{"Tg N"})

\item \code{EMISSIONS_CO()}: Emissions for carbon monoxide (\code{"Tg CO"})

\item \code{EMISSIONS_NMVOC()}: Emissions for non-methane volatile organic compounds (NMVOC) (\code{"Tg NMVOC"})

\item \code{EMISSIONS_OC()}: Emissions for organic carbon (\code{"Tg"})

\item \code{EMISSIONS_NH3()}: Emissions for NH3 (\code{"Tg"})

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers:
\code{\link[=CF4_CONSTRAIN]{CF4_CONSTRAIN()}},
\code{\link[=CONCENTRATIONS_CH4]{CONCENTRATIONS_CH4()}},
\code{\link[=CONCENTRATIONS_N2O]{CONCENTRATIONS_N2O()}},
\code{\link[=DELTA_CO2]{DELTA_CO2()}},
\code{\link[=EMISSIONS_CF4]{EMISSIONS_CF4()}},
\code{\link[=EMISSIONS_SO2]{EMISSIONS_SO2()}},
\code{\link[=FTOT_CONSTRAIN]{FTOT_CONSTRAIN()}},
\code{\link[=GLOBAL_TAS]{GLOBAL_TAS()}},
\code{\link[=NBP]{NBP()}},
\code{\link[=OCEAN_UPTAKE]{OCEAN_UPTAKE()}},
\code{\link[=RF_CF4]{RF_CF4()}},
\code{\link[=RF_TOTAL]{RF_TOTAL()}},
\code{\link[=RHO_BC]{RHO_BC()}},
\code{\link[=TRACKING_DATE]{TRACKING_DATE()}}
}
\concept{capability identifiers}
