% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{emissions}
\alias{emissions}
\alias{EMISSIONS_BC}
\alias{EMISSIONS_N2O}
\alias{EMISSIONS_NOX}
\alias{EMISSIONS_CO}
\alias{EMISSIONS_NMVOC}
\alias{EMISSIONS_OC}
\title{Identifiers for miscellaneous emissions not elsewhere described}
\usage{
EMISSIONS_BC()

EMISSIONS_N2O()

EMISSIONS_NOX()

EMISSIONS_CO()

EMISSIONS_NMVOC()

EMISSIONS_OC()
}
\description{
All of these variables can be set using the \code{\link{SETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{EMISSIONS_BC}: Black carbon emissions (\code{"Tg"})

\item \code{EMISSIONS_N2O}: N2O emissions (\code{"Tg N"})

\item \code{EMISSIONS_NOX}: Emissions for NOx compounds (\code{"Tg N"})

\item \code{EMISSIONS_CO}: Emissions for carbon monoxide (\code{"Tg CO"})

\item \code{EMISSIONS_NMVOC}: Emissions for non-methane volatile organic compounds (NMVOC) (\code{"Tg NMVOC"})

\item \code{EMISSIONS_OC}: Emissions for organic carbon (\code{"Tg"})
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{forcings}},
  \code{\link{haloemiss}}, \code{\link{haloforcings}},
  \code{\link{methane}}, \code{\link{ocean}},
  \code{\link{parameters}}, \code{\link{so2}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
