% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{fetchvars}
\alias{fetchvars}
\title{Fetch results from a running Hector core}
\usage{
fetchvars(core, dates, vars = NULL)
}
\arguments{
\item{core}{Hector core object}

\item{dates}{Vector of dates to fetch.  This will be automatically trimmed to
dates that are between the start date and the latest date currently run.}

\item{vars}{List (or vector) of capability strings defining the variables to
be fetched in the result.  By default, the list in the
\code{hector.default.fetchvars} option is used.}
}
\description{
This function will fetch results and return them as a data frame.  The core
must be active in order for this to work; attempting to fetch results from a
core that has been shut down will cause an error.
}
