% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{fetchvars_all}
\alias{fetchvars_all}
\title{Fetch all the available Hector variables}
\usage{
fetchvars_all(core, dates = NULL, scenario = NULL, write = F,
  outpath = NULL)
}
\arguments{
\item{core}{Hector core object}

\item{dates}{ector of dates, optional}

\item{scenario}{Scenario name, optional str}

\item{write}{If TRUE, write the resulting dataframe to a csv file, optional bool}

\item{outpath}{Absolute path of the output csv. Must be given if 'write' is true}
}
\value{
vars_all    Dataframe containing all Hector variables
}
\description{
This function is similar to fetchvars, except it fetches all available
Hector variables, including exogenous inputs (e.g., emissions & pre-industrial
concentrations), parameters (e.g., alpha, beta, ECS), and output
variables (e.g., concentrations & forcings).
}
\seealso{
Other main user interface functions: \code{\link{fetchvars}},
  \code{\link{newcore}}, \code{\link{reset}},
  \code{\link{run}}, \code{\link{setvar}},
  \code{\link{shutdown}}
}
\concept{main user interface functions}
