% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{forcings}
\alias{forcings}
\alias{RF_TOTAL}
\alias{RF_ALBEDO}
\alias{RF_CO2}
\alias{RF_H2O_STRAT}
\alias{RF_O3_TROP}
\alias{RF_VOL}
\alias{RF_MISC}
\alias{Q_CO2}
\alias{RF_CH4}
\title{Identifiers for capabilities in the Hector forcing component}
\usage{
RF_TOTAL()

RF_ALBEDO()

RF_CO2()

RF_H2O_STRAT()

RF_O3_TROP()

RF_VOL()

RF_MISC()

Q_CO2()

RF_CH4()
}
\description{
These identifiers specify forcing values that can be provided by hector via
the forcing component. All of the values corresponding to these identifiers
are read-only (\emph{i.e.}, they can only appear in \code{\link{GETDATA}}
messages.)
}
\section{Functions}{
\itemize{
\item \code{RF_TOTAL}: Total radiative forcing

\item \code{RF_ALBEDO}: Radiative forcing due to albedo

\item \code{RF_CO2}: Radiative forcing due to CO2

\item \code{RF_H2O_STRAT}: Radiative forcing due to stratospheric water vapor

\item \code{RF_O3_TROP}: Radiative forcing due to tropospheric ozone

\item \code{RF_VOL}: Radiative forcing due to volcanic activity

\item \code{RF_MISC}: Radiative miscellaneous forcing read in from ini.

\item \code{Q_CO2}: Radiative forcing efficiency for doubling of CO2

\item \code{RF_CH4}: Radiative forcing due to methane
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{rho} for parameters
related to radiative forcing efficiency.

Other capability identifiers: 
\code{\link{carboncycle}},
\code{\link{concentrations}},
\code{\link{constraints}},
\code{\link{delta}},
\code{\link{emissions}},
\code{\link{haloconstrain}},
\code{\link{haloemiss}},
\code{\link{haloforcings}},
\code{\link{methane}},
\code{\link{ocean}},
\code{\link{parameters}},
\code{\link{rho}},
\code{\link{so2}},
\code{\link{temperature}}
}
\concept{capability identifiers}
