% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_tracking_data}
\alias{get_tracking_data}
\title{Retrieve the tracking data for a Hector instance}
\usage{
get_tracking_data(core)
}
\arguments{
\item{core}{Handle to the Hector instance.}
}
\description{
Retrieve the tracking data for a Hector instance
}
