% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{get_tracking_data}
\alias{get_tracking_data}
\title{Retrieve the tracking data for a Hector instance}
\usage{
get_tracking_data(core)
}
\arguments{
\item{core}{Handle to the Hector instance.}
}
\value{
A \code{\link{data.frame}} with the tracking data. Columns include
\code{year} (integer), \code{pool_name} (character), \code{pool_value}
(double), \code{pool_units} (character), \code{source_name} (character),
and \code{source_fraction} (double).
}
\description{
Retrieve the tracking data for a Hector instance
}
\seealso{
Other main user interface functions: 
\code{\link{fetchvars}()},
\code{\link{newcore}()},
\code{\link{reset}()},
\code{\link{run}()},
\code{\link{setvar}()},
\code{\link{shutdown}()}
}
\concept{main user interface functions}
