% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{haloemiss}
\alias{haloemiss}
\alias{EMISSIONS_CF4}
\alias{EMISSIONS_C2F6}
\alias{EMISSIONS_HFC23}
\alias{EMISSIONS_HFC32}
\alias{EMISSIONS_HFC4310}
\alias{EMISSIONS_HFC125}
\alias{EMISSIONS_HFC134A}
\alias{EMISSIONS_HFC143A}
\alias{EMISSIONS_HFC227EA}
\alias{EMISSIONS_HFC245FA}
\alias{EMISSIONS_SF6}
\alias{EMISSIONS_CFC11}
\alias{EMISSIONS_CFC12}
\alias{EMISSIONS_CFC113}
\alias{EMISSIONS_CFC114}
\alias{EMISSIONS_CFC115}
\alias{EMISSIONS_CCL4}
\alias{EMISSIONS_CH3CCL3}
\alias{EMISSIONS_HCF22}
\alias{EMISSIONS_HCF141B}
\alias{EMISSIONS_HCF142B}
\alias{EMISSIONS_HALON1211}
\alias{EMISSIONS_HALON1301}
\alias{EMISSIONS_HALON2402}
\alias{EMISSIONS_CH3CL}
\alias{EMISSIONS_CH3BR}
\title{Identifiers for halocarbon emissions}
\usage{
EMISSIONS_CF4()

EMISSIONS_C2F6()

EMISSIONS_HFC23()

EMISSIONS_HFC32()

EMISSIONS_HFC4310()

EMISSIONS_HFC125()

EMISSIONS_HFC134A()

EMISSIONS_HFC143A()

EMISSIONS_HFC227EA()

EMISSIONS_HFC245FA()

EMISSIONS_SF6()

EMISSIONS_CFC11()

EMISSIONS_CFC12()

EMISSIONS_CFC113()

EMISSIONS_CFC114()

EMISSIONS_CFC115()

EMISSIONS_CCL4()

EMISSIONS_CH3CCL3()

EMISSIONS_HCF22()

EMISSIONS_HCF141B()

EMISSIONS_HCF142B()

EMISSIONS_HALON1211()

EMISSIONS_HALON1301()

EMISSIONS_HALON2402()

EMISSIONS_CH3CL()

EMISSIONS_CH3BR()
}
\description{
These identifiers correspond to emissions values for halocarbons.  They are
(for now) write only, meaning they can only appear in \code{\link{SETDATA}}
messages.  In all cases, the expected input units are gigagrams
(\code{"Gg"}).
}
\section{Functions}{
\itemize{
\item \code{EMISSIONS_CF4}: Emissions for CF4

\item \code{EMISSIONS_C2F6}: Emissions for C2F6

\item \code{EMISSIONS_HFC23}: Emissions for HFC-23

\item \code{EMISSIONS_HFC32}: Emissions for HFC-32

\item \code{EMISSIONS_HFC4310}: Emissions for HFC-4310

\item \code{EMISSIONS_HFC125}: Emissions for HFC-125

\item \code{EMISSIONS_HFC134A}: Emissions for HFC-134a

\item \code{EMISSIONS_HFC143A}: Emissions for HFC-143a

\item \code{EMISSIONS_HFC227EA}: Emissions for HFC-227ea

\item \code{EMISSIONS_HFC245FA}: Emissions for HFC-254fa

\item \code{EMISSIONS_SF6}: Emissions for sulfur hexafluoride

\item \code{EMISSIONS_CFC11}: Emissions for CFC-11

\item \code{EMISSIONS_CFC12}: Emissions for CFC-12

\item \code{EMISSIONS_CFC113}: Emissions for CFC-113

\item \code{EMISSIONS_CFC114}: Emissions for CFC-114

\item \code{EMISSIONS_CFC115}: Emissions for CFC-115

\item \code{EMISSIONS_CCL4}: Emissions for carbon tetrachloride

\item \code{EMISSIONS_CH3CCL3}: Emissions for trichloroethane

\item \code{EMISSIONS_HCF22}: Emissions for HCFC-22

\item \code{EMISSIONS_HCF141B}: Emissions for HCFC-141b

\item \code{EMISSIONS_HCF142B}: Emissions for HCFC-142b

\item \code{EMISSIONS_HALON1211}: Emissions for halon-1211

\item \code{EMISSIONS_HALON1301}: Emissions for halon-1301

\item \code{EMISSIONS_HALON2402}: Emissions for halon-2402

\item \code{EMISSIONS_CH3CL}: Emissions for chloromethane

\item \code{EMISSIONS_CH3BR}: Emissions for bromomethane
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{emissions}},
  \code{\link{forcings}}, \code{\link{haloforcings}},
  \code{\link{methane}}, \code{\link{ocean}},
  \code{\link{parameters}}, \code{\link{so2}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
