% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{hectorutil}
\alias{hectorutil}
\alias{isactive}
\alias{startdate}
\alias{enddate}
\alias{format.hcore}
\alias{print.hcore}
\title{Utility functions for Hector instances}
\usage{
isactive(core)

startdate(core)

enddate(core)

\method{format}{hcore}(core)

\method{print}{hcore}(core)
}
\arguments{
\item{core}{Hector instance to operate on}
}
\description{
Utility functions for Hector instances

\strong{isactive}: Indicate whether a Hector instance is active

\strong{startdate}: Report the start date for a Hector instance

\strong{enddate}: Report the end date for a Hector instance
}
