% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{loglevels}
\alias{loglevels}
\alias{LL_DEBUG}
\alias{LL_NOTICE}
\alias{LL_WARNING}
\alias{LL_SEVERE}
\title{Hector logging levels}
\usage{
LL_DEBUG()

LL_NOTICE()

LL_WARNING()

LL_SEVERE()
}
\description{
These functions return constants that can be used in
\code{\link{newcore}} to set the minimum logging level
}
\section{Functions}{
\itemize{
\item \code{LL_DEBUG}: Set logging at 'debug' level.

\item \code{LL_NOTICE}: Set logging at 'notice' level.

\item \code{LL_WARNING}: Set logging at 'warning' level.

\item \code{LL_SEVERE}: Set logging at 'severe' level.
}}

