% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{CONCENTRATIONS_CH4}
\alias{CONCENTRATIONS_CH4}
\alias{PREINDUSTRIAL_CH4}
\alias{EMISSIONS_CH4}
\alias{NATURAL_CH4}
\alias{LIFETIME_SOIL}
\alias{LIFETIME_STRAT}
\alias{methane}
\title{Identifiers for quantities in the methane component}
\usage{
CONCENTRATIONS_CH4()

PREINDUSTRIAL_CH4()

EMISSIONS_CH4()

NATURAL_CH4()

LIFETIME_SOIL()

LIFETIME_STRAT()
}
\description{
These identifiers correspond to variables that can be read and/or set in the
methane component.
}
\section{Output variables}{

These variables can be read from the methane component.
\describe{
\item{CONCENTRATIONS_CH4}{Atmospheric methane concentration}
\item{PREINDUSTRIAL_CH4}{Preindustrial methane concentration}
}
}

\section{Input variables}{

These variables can be set in the methane component. The expected units
string is given after each description.
\describe{
\item{EMISSIONS_CH4}{Methane emissions (\code{"Tg CH4"})}
\item{PREINDUSTRIAL_CH4}{Preindustrial methane concentration (\code{"ppbv CH4"})}
\item{CH4_CONSTRAIN}{N2O concentration constraint (\code{"ppbv CH4"})}
\item{NATURAL_CH4}{Natural methane emissions (\code{"Tg CH4"})}
\item{LIFETIME_SOIL}{Time scale for methane loss into soil (\code{"Years"})}
\item{LIFETIME_STRAT}{Time scale for methane loss into stratosphere (\code{"Years"})}
}
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_N2O}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{EMISSIONS_SO2}()},
\code{\link{FTOT_CONSTRAIN}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{NBP}()},
\code{\link{OCEAN_UPTAKE}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
