% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{methane}
\alias{methane}
\alias{ATMOSPHERIC_CH4}
\alias{PREINDUSTRIAL_CH4}
\alias{EMISSIONS_CH4}
\alias{NATURAL_CH4}
\alias{LIFETIME_SOIL}
\alias{LIFETIME_STRAT}
\title{Identifiers for quantities in the methane component}
\usage{
ATMOSPHERIC_CH4()

PREINDUSTRIAL_CH4()

EMISSIONS_CH4()

NATURAL_CH4()

LIFETIME_SOIL()

LIFETIME_STRAT()
}
\description{
These identifiers correspond to variables that can be read and/or set in the
methane component.
}
\section{Output variables}{

These variables can be read from the methane component.
\describe{
\item{ATMOSPHERIC_CH4}{Atmospheric methane concentration}
\item{PREINDUSTRIAL_CH4}{Preindustrial methane concentration}
}
}

\section{Input variables}{

These variables can be set in the methane component.  The expected units string is given after
each description.
\describe{
\item{EMISSIONS_CH4}{Methane emissions (\code{"Tg CH4"})}
\item{PREINDUSTRIAL_CH4}{Preindustrial methane concentration (\code{"ppbv CH4"})}
\item{NATURAL_CH4}{Natural methane emissions (\code{"Tg CH4"})}
\item{LIFETIME_SOIL}{Time scale for methane loss into soil (\code{"Years"})}
\item{LIFETIME_STRAT}{Time scale for methane loss into stratosphere (\code{"Years"})}
}
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{emissions}},
  \code{\link{forcings}}, \code{\link{haloemiss}},
  \code{\link{haloforcings}}, \code{\link{ocean}},
  \code{\link{parameters}}, \code{\link{so2}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
