% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newcore}
\alias{newcore}
\title{Create and initialize a new hector core}
\usage{
newcore(inifile)
}
\arguments{
\item{infile}{(String) name of the hector input file.}
}
\value{
handle for the Hector core.
}
\description{
The object returned is a handle to the newly created core.  It will be required as an
argument for all functions that operate on the core.  Creating multiple cores
simultaneously is supported.
}
