% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{OCEAN_UPTAKE}
\alias{OCEAN_UPTAKE}
\alias{OCEAN_C}
\alias{OCEAN_C_HL}
\alias{OCEAN_C_LL}
\alias{OCEAN_C_ML}
\alias{OCEAN_C_IO}
\alias{OCEAN_C_DO}
\alias{TT}
\alias{TU}
\alias{TWI}
\alias{TID}
\alias{PH_HL}
\alias{PH_LL}
\alias{PH}
\alias{HL_OCEAN_UPTAKE}
\alias{LL_OCEAN_UPTAKE}
\alias{PCO2_HL}
\alias{PCO2_LL}
\alias{PCO2}
\alias{DIC_HL}
\alias{DIC}
\alias{DIC_LL}
\alias{SST_HL}
\alias{SST_LL}
\alias{CO3_LL}
\alias{CO3_HL}
\alias{CO3}
\alias{ocean}
\title{Identifiers for quantities in the ocean component}
\usage{
OCEAN_UPTAKE()

OCEAN_C()

OCEAN_C_HL()

OCEAN_C_LL()

OCEAN_C_ML()

OCEAN_C_IO()

OCEAN_C_DO()

TT()

TU()

TWI()

TID()

PH_HL()

PH_LL()

PH()

HL_OCEAN_UPTAKE()

LL_OCEAN_UPTAKE()

PCO2_HL()

PCO2_LL()

PCO2()

DIC_HL()

DIC()

DIC_LL()

SST_HL()

SST_LL()

CO3_LL()

CO3_HL()

CO3()
}
\description{
These identifiers correspond to variables that can be read from the ocean
component using the \code{\link{GETDATA}} message type.
}
\section{Functions}{
\itemize{
\item \code{OCEAN_UPTAKE()}: Atmosphere-ocean carbon flux

\item \code{OCEAN_C()}: Ocean total carbon pool

\item \code{OCEAN_C_HL()}: Ocean surface high-latitude carbon pool

\item \code{OCEAN_C_LL()}: Ocean surface low-latitude carbon pool

\item \code{OCEAN_C_ML()}: Ocean surface or mixed layer carbon pool

\item \code{OCEAN_C_IO()}: Intermediate ocean carbon pool

\item \code{OCEAN_C_DO()}: Deep ocean carbon pool

\item \code{TT()}: Thermohaline overturning

\item \code{TU()}: High-latitude overturning

\item \code{TWI()}: Warm-intermediate exchange

\item \code{TID()}: Intermediate-deep exchange

\item \code{PH_HL()}: High-latitude pH

\item \code{PH_LL()}: Low-latitude pH

\item \code{PH()}: Ocean surface pH

\item \code{HL_OCEAN_UPTAKE()}: Atmosphere-ocean carbon flux, high-latitude

\item \code{LL_OCEAN_UPTAKE()}: Atmosphere-ocean carbon flux, low-latitude

\item \code{PCO2_HL()}: Partial pressure of CO2, high-latitude

\item \code{PCO2_LL()}: Partial pressure of CO2, low-latitude

\item \code{PCO2()}: Ocean surface partial pressure of CO2

\item \code{DIC_HL()}: Dissolved inorganic carbon, high-latitude

\item \code{DIC()}: Ocean surface dissolved inorganic carbon

\item \code{DIC_LL()}: Dissolved inorganic carbon, low-latitude

\item \code{SST_HL()}: Absolute ocean surface temperature, high-latitude (deg C)

\item \code{SST_LL()}: Absolute ocean surface temperature, low-latitude (deg C)

\item \code{CO3_LL()}: Carbonate concentration, low-latitude

\item \code{CO3_HL()}: Carbonate concentration, high-latitude

\item \code{CO3()}: Ocean surface carbonate concentration

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_CH4}()},
\code{\link{CONCENTRATIONS_N2O}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{EMISSIONS_SO2}()},
\code{\link{FTOT_CONSTRAIN}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{NBP}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
