% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{ocean}
\alias{ocean}
\alias{OCEAN_CFLUX}
\alias{OCEAN_C}
\alias{OCEAN_C_HL}
\alias{OCEAN_C_LL}
\alias{OCEAN_C_IO}
\alias{OCEAN_C_DO}
\alias{TT}
\alias{TU}
\alias{TWI}
\alias{TID}
\alias{PH_HL}
\alias{PH_LL}
\alias{ATM_OCEAN_FLUX_HL}
\alias{ATM_OCEAN_FLUX_LL}
\alias{PCO2_HL}
\alias{PCO2_LL}
\alias{DIC_HL}
\alias{DIC_LL}
\alias{TEMP_HL}
\alias{TEMP_LL}
\alias{CO3_LL}
\alias{CO3_HL}
\title{Identifiers for quantities in the ocean component}
\usage{
OCEAN_CFLUX()

OCEAN_C()

OCEAN_C_HL()

OCEAN_C_LL()

OCEAN_C_IO()

OCEAN_C_DO()

TT()

TU()

TWI()

TID()

PH_HL()

PH_LL()

ATM_OCEAN_FLUX_HL()

ATM_OCEAN_FLUX_LL()

PCO2_HL()

PCO2_LL()

DIC_HL()

DIC_LL()

TEMP_HL()

TEMP_LL()

CO3_LL()

CO3_HL()
}
\description{
These identifiers correspond to variables that can be read from the ocean
component using the \code{\link{GETDATA}} message type.
}
\section{Functions}{
\itemize{
\item \code{OCEAN_CFLUX}: Atmosphere-ocean carbon flux

\item \code{OCEAN_C}: Ocean total carbon pool

\item \code{OCEAN_C_HL}: Ocean surface high-latitude carbon pool

\item \code{OCEAN_C_LL}: Ocean surface low-latitude carbon pool

\item \code{OCEAN_C_IO}: Intermediate ocean carbon pool

\item \code{OCEAN_C_DO}: Deep ocean carbon pool

\item \code{TT}: Thermohaline overturning

\item \code{TU}: High-latitude overturning

\item \code{TWI}: Warm-intermediate exchange

\item \code{TID}: Intermediate-deep exchange

\item \code{PH_HL}: High-latitude Ph

\item \code{PH_LL}: Low-latitude Ph

\item \code{ATM_OCEAN_FLUX_HL}: Atmosphere-ocean carbon flux, high-latitude

\item \code{ATM_OCEAN_FLUX_LL}: Atmosphere-ocean carbon flux, low-latitude

\item \code{PCO2_HL}: Partial pressure of CO2, high-latitude

\item \code{PCO2_LL}: Partial pressure of CO2, low-latitude

\item \code{DIC_HL}: Dissolved inorganic carbon, high-latitude

\item \code{DIC_LL}: Dissolved inorganic carbon, low-latitude

\item \code{TEMP_HL}: Ocean temperature, high-latitude

\item \code{TEMP_LL}: Ocean temperature, low-latitude

\item \code{CO3_LL}: Carbonate concentration, low-latitude

\item \code{CO3_HL}: Carbonate concentration, high-latitude
}}

\section{To Do}{

Some of these could be made writeable as well as readable.  Notably, TT, TU,
TWI, and TID are only set from the input, not calculated.
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{emissions}},
  \code{\link{forcings}}, \code{\link{haloemiss}},
  \code{\link{haloforcings}}, \code{\link{methane}},
  \code{\link{parameters}}, \code{\link{so2}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
