% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reset}
\alias{reset}
\title{Reset a Hector instance to an earlier date}
\usage{
reset(core, date = 0)
}
\arguments{
\item{core}{Handle for the Hector instance that is to be reset.}

\item{date}{Date to reset to.  The default is to reset to the model start date with
a rerun of the spinup.}
}
\description{
Resetting the model returns it to its state at a previous time.  If the requested time
is before the model start date (any value will do; conventionally zero is used), then
the spinup will be rerun, and the model will be
left ready to run at the start date.  (By contrast, resetting \emph{to} the start
date leaves the model ready to run at the start date, but without having rerun the
spinup.)
}
\seealso{
Other main user interface functions: \code{\link{fetchvars}},
  \code{\link{newcore}}, \code{\link{run}},
  \code{\link{setvar}}, \code{\link{shutdown}}
}
\concept{main user interface functions}
