% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{rho}
\alias{rho}
\alias{RHO_SO2}
\title{Identifiers for Hector forcing component parameters}
\usage{
RHO_SO2()
}
\description{
These identifiers specify radiative forcing efficiency in hector these values
can be read and/or set by hectors forcing and halocarbon components.
the forcing component.
}
\section{Functions}{
\itemize{
\item \code{RHO_SO2()}: a radiative forcing efficiency for SO2

}}
\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{forcings}
forcing values provided from the hector forcing component.

Other capability identifiers: 
\code{\link{carboncycle}},
\code{\link{concentrations}},
\code{\link{constraints}},
\code{\link{delta}},
\code{\link{emissions}},
\code{\link{forcings}},
\code{\link{haloconstrain}},
\code{\link{haloemiss}},
\code{\link{haloforcings}},
\code{\link{methane}},
\code{\link{ocean}},
\code{\link{parameters}},
\code{\link{so2}},
\code{\link{temperature}}
}
\concept{capability identifiers}
