% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{rho}
\alias{rho}
\alias{rho_CO2}
\alias{rho_N2O}
\alias{rho_CH4}
\alias{rho_tropO3}
\alias{rho_SO2d}
\alias{rho_SO2i}
\title{Identifiers for Hector forcing component parameters}
\usage{
rho_CO2()

rho_N2O()

rho_CH4()

rho_tropO3()

rho_SO2d()

rho_SO2i()
}
\description{
These identifiers specify radiative forcing efficiency in hector these values
can be read and/or set by hectors forcing and halocarbon components.
the forcing component.
}
\section{Functions}{
\itemize{
\item \code{rho_CO2}: CO2 radiative forcing per doubling of CO2

\item \code{rho_N2O}: Radiative forcing efficiency for N2O

\item \code{rho_CH4}: Radiative forcing efficiency for CH4

\item \code{rho_tropO3}: Radiative forcing efficiency for tropospheric O3

\item \code{rho_SO2d}: Direct forcing per unit sulphate aerosols emissions

\item \code{rho_SO2i}: Indirect forcing per unit sulphate aerosols emissions
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
\link{haloforcings} for forcings from halocarbons and \link{forcings}
forcing values provided from the hector forcing component.

Other capability identifiers: 
\code{\link{carboncycle}},
\code{\link{concentrations}},
\code{\link{constraints}},
\code{\link{emissions}},
\code{\link{forcings}},
\code{\link{haloconstrain}},
\code{\link{haloemiss}},
\code{\link{haloforcings}},
\code{\link{methane}},
\code{\link{ocean}},
\code{\link{parameters}},
\code{\link{so2}},
\code{\link{temperature}}
}
\concept{capability identifiers}
