% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Run the Hector climate model}
\usage{
run(core, runtodate = -1)
}
\arguments{
\item{core}{Handle a Hector instance that is to be run.}

\item{runtodate}{Date to run to.  The default is to run to the end date configured
in the input file used to initialize the core.}
}
\description{
Run Hector up through the specified time.  This function does not return the results
of the run.  To get results, run \code{fetch}.
}
