% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shutdown}
\alias{shutdown}
\title{Shutdown a hector core}
\usage{
shutdown(core)
}
\arguments{
\item{core}{Handle to the core that is to be shut down.}
}
\value{
The core handle, modified to show that it is no longer active.
}
\description{
Shutting down a core will free the core itself and all of the objects it created. Any attempted
operation on the core after that will raise an error.
}
