% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shutdown}
\alias{shutdown}
\title{Shutdown a hector instance}
\usage{
shutdown(core)
}
\arguments{
\item{core}{Handle to a Hector instance}
}
\value{
The Hector instance handle
}
\description{
Shutting down an instance will free the instance itself and all of the objects it created. Any attempted
operation on the instance after that will raise an error.
}
\section{Caution}{

This function should be called as \code{mycore <- shutdown(mycore)} so that the change
from active to inactive will be recorded in the caller.
}

\seealso{
Other main user interface functions: \code{\link{fetchvars}},
  \code{\link{newcore}}, \code{\link{reset}},
  \code{\link{run}}, \code{\link{setvar}}
}
\concept{main user interface functions}
