% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aadoc.R
\name{EMISSIONS_SO2}
\alias{EMISSIONS_SO2}
\alias{VOLCANIC_SO2}
\alias{so2}
\title{Identifiers for quantities in the SO2 component}
\usage{
EMISSIONS_SO2()

VOLCANIC_SO2()
}
\description{
These identifiers correspond to variables that can be read and/or set in the
SO2 component
}
\section{Output variables}{

These variables can be read using the \code{\link{GETDATA}} message type.
\describe{
\item{NATURAL_SO2}{Natural SO2 emissions}
\item{Y2000_SO2}{Year 2000 SO2 emissions}
\item{EMISIONS_SO2}{Anthropogenic SO2 emissions}
\item{VOLCANIC_SO2}{Forcing due to volcanic SO2 emissions}
}
}

\section{Input variables}{

These variables can be set using the \code{\link{SETDATA}} message type.
\describe{
\item{EMISSIONS_SO2}{Anthropogenic SO2 emissions (\code{"Gg S"})}
\item{VOLCANIC_SO2}{Forcing due to volcanic SO2 emissions (\code{"W/m2"})}
}
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as functions. Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: 
\code{\link{CF4_CONSTRAIN}()},
\code{\link{CONCENTRATIONS_CH4}()},
\code{\link{CONCENTRATIONS_N2O}()},
\code{\link{DELTA_CO2}()},
\code{\link{EMISSIONS_BC}()},
\code{\link{EMISSIONS_CF4}()},
\code{\link{FTOT_CONSTRAIN}()},
\code{\link{GLOBAL_TAS}()},
\code{\link{NBP}()},
\code{\link{OCEAN_UPTAKE}()},
\code{\link{RF_CF4}()},
\code{\link{RF_TOTAL}()},
\code{\link{RHO_BC}()},
\code{\link{TRACKING_DATE}()}
}
\concept{capability identifiers}
