% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{so2}
\alias{so2}
\alias{NATURAL_SO2}
\alias{Y2000_SO2}
\alias{EMISSIONS_SO2}
\alias{VOLCANIC_SO2}
\title{Identifiers for quantities in the SO2 component}
\usage{
NATURAL_SO2()

Y2000_SO2()

EMISSIONS_SO2()

VOLCANIC_SO2()
}
\description{
These identifiers correspond to variables that can be read and/or set in the
SO2 component
}
\section{Output variables}{

These variables can be read using the \code{\link{GETDATA}} message type.
\describe{
\item{NATURAL_SO2}{Natural SO2 emissions}
\item{Y2000_SO2}{Year 2000 SO2 emissions}
\item{EMISIONS_SO2}{Anthropogenic SO2 emissions}
\item{VOLCANIC_SO2}{Forcing due to volcanic SO2 emissions}
}
}

\section{Input variables}{

These variables can be set using the \code{\link{SETDATA}} message type.
\describe{
\item{EMISSIONS_SO2}{Anthropogenic SO2 emissions (\code{"Gg S"})}
\item{VOLCANIC_SO2}{Forcing due to volcanic SO2 emissions (\code{"W/m2"})}
}
}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{emissions}},
  \code{\link{forcings}}, \code{\link{haloemiss}},
  \code{\link{haloforcings}}, \code{\link{methane}},
  \code{\link{ocean}}, \code{\link{parameters}},
  \code{\link{temperature}}
}
\concept{capability identifiers}
