% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{split_biome}
\alias{split_biome}
\title{Create new biomes by splitting up an existing biome}
\usage{
split_biome(core, old_biome, new_biomes,
  fveg_c = rep(1/length(new_biomes), length(new_biomes)),
  fdetritus_c = fveg_c, fsoil_c = fveg_c, fnpp_flux0 = fveg_c, ...)
}
\arguments{
\item{core}{Hector core}

\item{old_biome}{Name of biome that will be split up}

\item{new_biomes}{Names of biomes to be created}

\item{fveg_c}{Fractions of vegetation C distributed to each biome
in `new_biomes`. Must be the same length as `new_biomes`. Default
is to split vegetation C evenly between all biomes.}

\item{fdetritus_c}{Fractions of detritus C distributed to each
biome. Defaults to the same value as `fveg_c`.}

\item{fsoil_c}{Fractions of soil C distributed to each biome.
Defaults to the same value as `fveg_c`.}

\item{fnpp_flux0}{Fraction of initial NPP flux distributed to each
biome. Defaults to the same value as `fveg_c`.}

\item{...}{Additional biome-specific parameters, as set by
[create_biome()]. Note that these are passed to [create_biome()]
via [base::mapply()], so they can be vectorized across biomes.}
}
\description{
Distributes vegetation, detritus, and soil C, and initial NPP flux
from an existing biome across multiple new biomes.
}
