% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R, R/RcppExports.R
\name{temperature}
\alias{temperature}
\alias{GLOBAL_TEMP}
\alias{GLOBAL_TEMPEQ}
\alias{OCEAN_SURFACE_TEMP}
\alias{OCEAN_AIR_TEMP}
\alias{LAND_AIR_TEMP}
\alias{FLUX_MIXED}
\alias{FLUX_INTERIOR}
\alias{HEAT_FLUX}
\title{Identifiers for variables associated with the temperature component}
\usage{
GLOBAL_TEMP()

GLOBAL_TEMPEQ()

OCEAN_SURFACE_TEMP()

OCEAN_AIR_TEMP()

LAND_AIR_TEMP()

FLUX_MIXED()

FLUX_INTERIOR()

HEAT_FLUX()
}
\description{
All of these variables can be read using the \code{\link{GETDATA}} message
type.
}
\section{Functions}{
\itemize{
\item \code{GLOBAL_TEMP}: Global mean temperature

\item \code{GLOBAL_TEMPEQ}: Equilibrium global temperature

\item \code{OCEAN_SURFACE_TEMP}: Average ocean surface temperature anomaly

\item \code{OCEAN_AIR_TEMP}: Average ocean air temperature anomaly

\item \code{LAND_AIR_TEMP}: Average land temperature anomaly

\item \code{FLUX_MIXED}: Heat flux into the mixed layer of the ocean

\item \code{FLUX_INTERIOR}: Heat flux into the interior layer of the ocean

\item \code{HEAT_FLUX}: Total heat flux into the ocean
}}

\section{Note}{

Because these identifiers are provided as \code{#define} macros in the hector code,
these identifiers are provided in the R interface as function.  Therefore,
these objects must be called to use them; \emph{e.g.}, \code{GETDATA()}
instead of the more natural looking \code{GETDATA}.
}

\seealso{
Other capability identifiers: \code{\link{carboncycle}},
  \code{\link{concentrations}}, \code{\link{emissions}},
  \code{\link{forcings}}, \code{\link{haloemiss}},
  \code{\link{haloforcings}}, \code{\link{methane}},
  \code{\link{ocean}}, \code{\link{parameters}},
  \code{\link{so2}}
}
\concept{capability identifiers}
